refactor: move stripQuotes util
